refactor(admin): extract block-form picker-dialog module
Verbatim move of the singleton media picker into
block-form/picker-dialog.ts: dialog mount/close (including the
module-load DOMContentLoaded mount, preserved through the import chain),
per-open state, grid rendering, pagination, upload flow and the
selectPickerImage/File write-back. openPickerDialog is the module's
only export.

Two source-guard tests pinned the moved code by path and had to follow
it now rather than in the facade slice: admin-upload-error-feedback
repoints its picker-upload slice to picker-dialog.ts, and
block-form-canonical-escape now scans the facade plus every
block-form/ module (comment-stripped) instead of one hardcoded file,
so the remaining slices cannot break it again.

Part of #38, slice 5/6.
